refactor(rolls): extract client-roll module
refs #559
